+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * [Added] metadata check [#91](https://github.com/dalance/veryl/issues/91)
 * [Added] dependency support [#36](https://github.com/dalance/veryl/issues/36)
 * [Added] veryl update command
+* [Added] language server metadata support [#95](https://github.com/dalance/veryl/issues/95)
 
 ## [v0.2.1](https://github.com/dalance/veryl/compare/v0.2.0...v0.2.1) - 2023-01-19